Crate dynlib

DynLib

Provides a relatively safe interface to interact with windows dll loading. This crate should only be used on windows platforms. If you attempt to use it on non-windows platform you will likely get a compiler error.

The entry point to this crate is LoadWinDynLib. This allows the developer to set configuration flags. There are some details contained within the type page. Most specifcally one should check which flags are and are not allowed. There is some compatibility issues between flags, as they'll modify where files are located.

The loaded library does not have a built in drop flag. You are responsible for dropping your loaded DLL's. Generally speaking you have to be responsible for this. When the DLL is unloaded from memory the symbols/functions you have loaded are invalidated, attempts to call them will result in a memory fault, and likely process termination.

An example project to generate a win dll is here]. A few cargo directives are required, as well you must ensure you are building with the rust-msvc compiler.

Structs

DynLibWin
DynLib

LoadWinDynLib
Options for loading a Dynlib/Exe into windows.

Type Definitions

VoidPtr
Untyped Pointer to a Function.